+++ tests/test_persistence_models.py
@@ -20,7 +20,8 @@ class Config:
 
 @yorm.attr(key=yorm.types.String)
 @yorm.attr(name=yorm.types.String)
-@yorm.sync("{self.root}/{self.key}/config.yml")
+@yorm.sync("{self.root}/{self.key}/config.yml",
+           auto_create=False, auto_save=False)
 class ConfigModel:
 
     def __init__(self, key, root):
@@ -41,21 +42,9 @@ class ConfigStore:
 
     def __init__(self, root):
         self.root = root
-        self.path = self.root + "/{}/config.yml"
-        self.attrs = dict(key=yorm.types.String,
-                          name=yorm.types.String)
 
     def read(self, key):
-        instance = Config(key)
-        path = self.path.format(key)
-        attrs = self.attrs
-        try:
-            yorm.sync(instance, path, attrs, existing=True, auto=False)
-        except yorm.exceptions.FileMissingError:
-            return None
-        else:
-            yorm.update_object(instance)
-            return instance
+        return yorm.find(ConfigModel, self.root, key)
 
 
 # TESTS ########################################################################
@@ -85,7 +74,6 @@ class TestPersistanceMapping:  # pylint: disable=no-member
     def test_nonmapped_attribute_is_kept(self):
         model = ConfigModel('my_key', self.root)
         model.unmapped = 42
-        yorm.update(model, force=True)
         assert 42 == model.unmapped