+++ yorm/test/test_bases_mappable.py
@@ -174,7 +174,7 @@ class TestMappable:
 
     def test_new(self):
         """Verify new attributes are added to the object."""
-        self.sample.__mapper__.strict = False
+        self.sample.__mapper__.auto_attr = True
         text = strip("""
         new: 42
         """)
@@ -183,7 +183,7 @@ class TestMappable:
 
     def test_new_unknown(self):
         """Verify an exception is raised on new attributes w/ unknown types"""
-        self.sample.__mapper__.strict = False
+        self.sample.__mapper__.auto_attr = True
         text = strip("""
         new: !!timestamp 2001-12-15T02:59:43.1Z
         """)
@@ -214,7 +214,7 @@ class TestMappableTriggers:
         self.sample = self.Sample()
         self.sample.__mapper__.fetch.reset_mock()
         self.sample.__mapper__.store.reset_mock()
-        self.sample.__mapper__.store_after_fetch = False
+        self.sample.__mapper__.auto_store_after_fetch = False
 
     def test_getattribute(self):
         with pytest.raises(AttributeError):
@@ -257,7 +257,7 @@ class TestMappableTriggers:
         self.sample.append('bar')
         self.sample.__mapper__.fetch.reset_mock()
         self.sample.__mapper__.store.reset_mock()
-        self.sample.__mapper__.store_after_fetch = False
+        self.sample.__mapper__.auto_store_after_fetch = False
         self.sample.__mapper__.modified = True
 
         for item in self.sample: