+++ yorm/test/test_decorators.py
@@ -7,7 +7,6 @@ from unittest.mock import patch, Mock
 import pytest
 from expecter import expect
 
-from yorm import exceptions
 from yorm import decorators
 from yorm.bases import Converter
 
@@ -55,27 +54,15 @@ class TestSyncObject:
     def test_multiple(self):
         """Verify mapping cannot be enabled twice."""
         sample = decorators.sync(self.Sample(), "sample.yml")
-        with pytest.raises(exceptions.MappingError):
+        with pytest.raises(TypeError):
             decorators.sync(sample, "sample.yml")
 
     @patch('yorm.diskutils.exists', Mock(return_value=True))
     def test_init_existing(self):
         """Verify an existing file is read."""
         with patch('yorm.diskutils.read', Mock(return_value="abc: 123")):
-            sample = decorators.sync(self.Sample(), "sample.yml", strict=False)
+            sample = decorators.sync(self.Sample(), "s.yml", auto_attr=True)
         assert 123 == sample.abc
-
-    @patch('yorm.diskutils.exists', Mock(return_value=False))
-    def test_exception_when_file_expected_but_missing(self):
-        decorators.sync(self.Sample(), "sample.yml", existing=False)
-        with pytest.raises(exceptions.FileMissingError):
-            decorators.sync(self.Sample(), "sample.yml", existing=True)
-
-    @patch('yorm.diskutils.exists', Mock(return_value=True))
-    def test_exception_when_file_not_expected_but_found(self):
-        decorators.sync(self.Sample(), "sample.yml", existing=True)
-        with pytest.raises(exceptions.FileAlreadyExistsError):
-            decorators.sync(self.Sample(), "sample.yml", existing=False)
 
 
 @patch('yorm.diskutils.write', Mock())
@@ -84,7 +71,7 @@ class TestSyncObject:
 class TestSyncInstances:
     """Unit tests for the `sync_instances` decorator."""
 
-    @decorators.sync("sample.yml", strict=False)
+    @decorators.sync("sample.yml", auto_attr=True)
     class SampleDecorated:
         """Sample decorated class using a single path."""
 
@@ -134,10 +121,6 @@ class TestSyncInstances:
     class SampleDecoratedWithAttributes:
         """Sample decorated class using a single path."""
 
-    @decorators.sync("sample.yml", attrs={'var1': MockConverter}, auto=False)
-    class SampleDecoratedWithAttributesAutoOff:
-        """Sample decorated class using a single path."""
-
     def test_no_attrs(self):
         """Verify mapping can be enabled with no attributes."""
         sample = self.SampleDecorated()
@@ -172,7 +155,7 @@ class TestSyncInstances:
         assert "path/to/two.yml" == sample2.__mapper__.path
 
     def test_filename_attributes_automatic(self):
-        """Verify attributes can be used to determine filename (auto)."""
+        """Verify attributes can be used to determine filename (auto save)."""
         sample1 = self.SampleDecoratedAttributesAutomatic('one')
         sample2 = self.SampleDecoratedAttributesAutomatic('two')
         assert "path/to/one.yml" == sample1.__mapper__.path