+++ yorm/types/containers.py
@@ -19,7 +19,7 @@ class Dictionary(Container, dict):
     @classmethod
     def to_data(cls, value):
         value2 = cls.create_default()
-        value2.update_value(value, strict=True)
+        value2.update_value(value, auto_attr=False)
 
         data = common.attrs[cls].__class__()
         for name, converter in common.attrs[cls].items():
@@ -27,7 +27,7 @@ class Dictionary(Container, dict):
 
         return data
 
-    def update_value(self, data, strict=False):
+    def update_value(self, data, *, auto_attr=True):
         cls = self.__class__
         value = cls.create_default()
 
@@ -50,13 +50,13 @@ class Dictionary(Container, dict):
             try:
                 converter = attrs.pop(name)
             except KeyError:
-                if strict:
+                if auto_attr:
+                    converter = standard.match(name, data2, nested=True)
+                    common.attrs[cls][name] = converter
+                else:
                     msg = "Ignored unknown nested file attribute: %s = %r"
                     log.warning(msg, name, data2)
                     continue
-                else:
-                    converter = standard.match(name, data2, nested=True)
-                    common.attrs[cls][name] = converter
 
             try:
                 attr = self[name]
@@ -65,7 +65,7 @@ class Dictionary(Container, dict):
 
             if all((isinstance(attr, converter),
                     issubclass(converter, Container))):
-                attr.update_value(data2, strict=strict)
+                attr.update_value(data2, auto_attr=auto_attr)
             else:
                 attr = converter.to_value(data2)
 
@@ -107,7 +107,7 @@ class List(Container, list):
     @classmethod
     def to_data(cls, value):
         value2 = cls.create_default()
-        value2.update_value(value, strict=True)
+        value2.update_value(value, auto_attr=False)
 
         data = []
 
@@ -117,7 +117,7 @@ class List(Container, list):
 
         return data
 
-    def update_value(self, data, strict=False):
+    def update_value(self, data, *, auto_attr=True):
         cls = self.__class__
         value = cls.create_default()
 
@@ -134,7 +134,7 @@ class List(Container, list):
 
             if all((isinstance(attr, converter),
                     issubclass(converter, Container))):
-                attr.update_value(item, strict=strict)
+                attr.update_value(item, auto_attr=auto_attr)
             else:
                 attr = converter.to_value(item)  # pylint: disable=no-member